refactor: stragegy package 추가
