+++ src/main/java/step2/domain/Lotto.java
@@ -1,9 +1,10 @@
 package step2.domain;
 
+import step2.domain.strategy.LottoPickStrategy;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class Lotto {
     private final List<LottoNumber> pick = new ArrayList<>();
